+++ src/main/java/chess/domain/piece/info/Diagonal.java
@@ -1,7 +1,6 @@
-package chess.domain;
+package chess.domain.piece.info;
 
 import chess.domain.piece.CurrentPieces;
-import chess.domain.piece.Position;
 
 import java.util.Arrays;
 import java.util.function.BiPredicate;
