+++ src/main/java/chess/domain/piece/info/Position.java
@@ -1,4 +1,4 @@
-package chess.domain.piece;
+package chess.domain.piece.info;
 
 import java.util.*;
