+++ src/test/java/chess/domain/piece/EmptyTest.java
@@ -1,5 +1,7 @@
 package chess.domain.piece;
 
+import chess.domain.piece.info.Color;
+import chess.domain.piece.info.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
