+++ src/test/java/chess/domain/piece/info/ColorTest.java
@@ -1,5 +1,6 @@
-package chess.domain.piece;
+package chess.domain.piece.info;
 
+import chess.domain.piece.info.Color;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
