+++ src/test/java/chess/domain/piece/info/ScoreTest.java
@@ -1,5 +1,6 @@
-package chess.domain.piece;
+package chess.domain.piece.info;
 
+import chess.domain.piece.info.Score;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
